Upgrade/Install: Move location of `$upgrade_notice` for better consistency.
Since the `$auto_upgrade_notice` has been added to the `update-core.php` table listing if an `$upgrade_notice` is present by virtue of the `readme.txt` 
containing data in the `== Upgrade Notice ==` section, this text will appear in between the compatibility information and the auto update information.

This changeset provides a more consistent appearance in the interface.

Props afragen, costdev, zunaid321.
Fixes #57939.




git-svn-id: https://develop.svn.wordpress.org/trunk@56001 602fd350-edb4-49c9-b593-d223f7449a82